revert(modbus): take the I/O image sizing out of this demand
DOPE-442. `io` / `ioMax` in the VPP manifest and the whole `io_sizes.h` chain
come out. DOPE-615 owns sizing the I/O image from the project, and lists
"reintroducing a per device image size in the VPP manifest" as out of scope --
so shipping both would put two demands on one surface, with this one declaring
per-device numbers that the other is designed to derive.

What goes:

  * the manifest's `device.io` (76 devices) and `device.ioMax` (57), their
    schema definitions, and every hop that carried them: board-info-resolver,
    hardware-module, resolve-board-selection, the pipeline inputs, BoardInfo;
  * `generate-io-sizes.ts` and the emission of `src/io_sizes.h`;
  * the profile's `derivedCounts` / `minCounts` / `maxCounts` / `countsSource`
    on baremetal, and the amber caveat the screen printed when a package
    declared no sizes.

What stays, deliberately: the Runtime v4 half of the buffer-mapping block. Its
counts come from the server's own `bufferMapping` and reach the device through
`conf/modbus_slave.json`, so they are neither derived from a manifest nor
anything DOPE-615 touches.

The firmware files go back to `development` untouched. The `#ifndef MAX_*`
guards, the `__has_include("io_sizes.h")` hook and the uint8_t -> uint16_t
widening of the register store all came from one commit that belongs to
DOPE-370, and openplc-editor#1079 is open with the same three files -- keeping a
second copy here would only conflict at merge.

One thing this restores rather than removes: with no sizes to report, the
screen's "This target reports no Modbus buffer sizes, so no address map can be
shown" branch is reachable again. That is what the demand's AC04 specified in
the first place.

+++ resources/sources/Baremetal/modbus_registers.cpp
@@ -9,7 +9,7 @@ Copyright (C) 2022 OpenPLC - Thiago Alves
 // In a debug-only build this whole TU compiles to nothing, saving flash/SRAM.
 #ifdef MODBUS_ENABLED
 
-bool init_mbregs(uint16_t size_holding, uint16_t size_dint_memory, uint16_t size_lint_memory, uint16_t size_coils, uint16_t size_inputregs, uint16_t size_inputstatus)
+bool init_mbregs(uint8_t size_holding, uint8_t size_dint_memory, uint8_t size_lint_memory, uint8_t size_coils, uint8_t size_inputregs, uint8_t size_inputstatus)
 {
     //Save sizes
     modbus.holding_size = size_holding;
@@ -64,9 +64,7 @@ bool init_mbregs(uint16_t size_holding, uint16_t size_dint_memory, uint16_t size
 
 bool get_discrete(uint16_t addr, bool regtype)
 {
-    // byte_addr indexes the packed bit bank, so it is addr/8 of a 16-bit
-    // address -- it does not fit in a uint8_t past 2040 bits.
-    uint16_t byte_addr = addr / 8;
+    uint8_t byte_addr = addr / 8;
     uint8_t bit_addr = addr % 8;
     if (regtype == COILS)
         return bitRead(modbus.coils[byte_addr], bit_addr);
@@ -76,7 +74,7 @@ bool get_discrete(uint16_t addr, bool regtype)
 
 void write_discrete(uint16_t addr, bool regtype, bool value)
 {
-    uint16_t byte_addr = addr / 8;
+    uint8_t byte_addr = addr / 8;
     uint8_t bit_addr = addr % 8;
     if (regtype == COILS)
         bitWrite(modbus.coils[byte_addr], bit_addr, value);
@@ -118,7 +116,7 @@ void readRegisters(uint16_t startreg, uint16_t numregs)
 
     uint16_t val;
     uint16_t i = 0;
-    uint16_t pos = 0;
+    uint8_t pos = 0;
 	while(numregs--)
     {
         if ((startreg + i) < modbus.holding_size)
@@ -179,7 +177,7 @@ void writeSingleRegister(uint16_t reg, uint16_t value)
         return;
     }
 
-    uint16_t pos = 0;
+    uint8_t pos = 0;
 
     if (reg < modbus.holding_size)
     {
@@ -256,7 +254,7 @@ void writeMultipleRegisters(uint16_t startreg, uint16_t numoutputs, uint8_t byte
 
     uint16_t value;
     uint16_t i = 0;
-    uint16_t pos = 0;
+    uint8_t pos = 0;
 	while(numoutputs--)
     {
         value = (uint16_t)mb_frame[7+i*2] << 8 | (uint16_t)mb_frame[8+i*2];
